Decompose the 445-line WebRoutes.index dashboard handler (#1120)
The two heavy phases move to WebRoutes+IndexRows.swift: the per-student
grade-data load becomes loadStudentDashboardGradeData (returning a typed
DashboardGradeData) and the per-setup row build becomes buildTestSetupRow
(reading a typed IndexRowContext), leaving index as state dispatch plus
the load/sort/group pipeline. The swiftlint function_body_length /
cyclomatic_complexity suppression on index is removed. This commit also
carries the dashboard's adoption of the #1118 shared display-order
comparator and section-grouping fold (same file).

No behavior change.
