+++ Sources/APIServer/Routes/Web/WebRoutes.swift
@@ -36,13 +36,11 @@ struct WebRoutes: RouteCollection {
 
     // The student dashboard handler dispatches across many distinct
     // page states: no active course → /enroll, instructor-only courses
-    // → admin index, no assignments yet → empty-course page, etc.  Each
-    // arm builds a different Leaf context with disjoint data needs, so
-    // splitting them would require either a state-machine wrapper or
-    // duplicating the auth + course resolution preamble at each arm's
-    // entry point.  Both are noisier than the inline switch.
+    // → admin index, no assignments yet → empty-course page, etc.  The
+    // heavy phases (per-student grade data, per-setup row build) live in
+    // WebRoutes+IndexRows.swift (#1120); what remains here is the state
+    // dispatch plus the load/sort/group pipeline.
     @Sendable
-    // swiftlint:disable:next function_body_length cyclomatic_complexity
     func index(req: Request) async throws -> Response {
         let user = try req.auth.require(APIUser.self)
 
@@ -153,163 +151,17 @@ struct WebRoutes: RouteCollection {
                 .all()
         }
 
-        var latestSubmissionBySetupID: [String: LatestSubmissionItem] = [:]
-        var submissionCountBySetupID: [String: Int] = [:]
-        var bestGradePercentBySetupID: [String: Int] = [:]
-        var overridePercentBySetupID: [String: Int] = [:]
-        var latestBadgesBySetupID: [String: [AchievementBadge]] = [:]
-        if let userID = user.id {
-            let setupIDs = setups.compactMap(\.id)
-            if !setupIDs.isEmpty {
-                // Instructor grade overrides for this student take precedence
-                // over the runner-computed best grade below.  The override and
-                // submission reads are independent and run concurrently.
-                async let overridesFetch = loadGradeOverridePercents(setupIDs: setupIDs, on: req.db)
-                async let submissionsFetch = APISubmission.query(on: req.db)
-                    .filter(\.$userID == userID)
-                    .filter(\.$testSetupID ~~ setupIDs)
-                    .filter(\.$kind == APISubmission.Kind.student)
-                    .sort(\.$submittedAt, .descending)
-                    .all()
-                let overrideMap = try await overridesFetch
-                for setupID in setupIDs {
-                    if let pct = overrideMap[GradeOverrideKey(setupID: setupID, userID: userID)] {
-                        overridePercentBySetupID[setupID] = pct
-                    }
-                }
-                let submissions = try await submissionsFetch
-
-                var grouped: [String: [APISubmission]] = [:]
-                for submission in submissions {
-                    grouped[submission.testSetupID, default: []].append(submission)
-                }
-
-                for (setupID, items) in grouped {
-                    submissionCountBySetupID[setupID] = items.count
-                    if let latest = items.first {
-                        let when = latest.submittedAt.map { fmt.string(from: $0) } ?? "—"
-                        latestSubmissionBySetupID[setupID] = LatestSubmissionItem(
-                            submissionID: latest.id ?? "",
-                            submittedAtText: when
-                        )
-                    }
-                }
-
-                let submissionIDs = submissions.compactMap(\.id)
-                if !submissionIDs.isEmpty {
-                    // Results plus the three achievement reads only share
-                    // inputs computed above, so all four run concurrently.
-                    async let resultsFetch = APIResult.query(on: req.db)
-                        .filter(\.$submissionID ~~ submissionIDs)
-                        .sort(\.$receivedAt, .descending)
-                        .all()
-                    async let disabledFetch = BuiltInAchievements.disabledBySetup(
-                        setupIDs: Array(setupIDs), on: req.db)
-                    async let perSubFetch = BuiltInAchievements.manifestPerSubmissionBySetup(
-                        setupIDs: Array(setupIDs), on: req.db)
-                    // Class-wide badges this user currently holds across all setups.
-                    async let classAchievementsFetch = APIClassAchievement.query(on: req.db)
-                        .filter(\.$userID == userID)
-                        .filter(\.$testSetupID ~~ setupIDs)
-                        .all()
-                    let resultRows = try await resultsFetch
-
-                    // Best grade percentage per submission across ALL result
-                    // sources — the shared "highest grade wins" fold applied
-                    // to the rows we already fetched (#1111).
-                    var resultsBySubmissionID: [String: [APIResult]] = [:]
-                    for row in resultRows {
-                        resultsBySubmissionID[row.submissionID, default: []].append(row)
-                    }
-                    let bestPercentBySubmissionID =
-                        resultsBySubmissionID
-                        .compactMapValues { bestGradePercent(of: $0) }
-                    // One preferred result per submission (worker-first) is still
-                    // needed for achievement-badge display below.
-                    var preferredResultBySubmissionID: [String: APIResult] = [:]
-                    for row in resultRows {
-                        let key = row.submissionID
-                        if let existing = preferredResultBySubmissionID[key] {
-                            let existingSource = existing.source ?? "worker"
-                            let currentSource = row.source ?? "worker"
-                            if existingSource == "worker" { continue }
-                            if currentSource == "worker" {
-                                preferredResultBySubmissionID[key] = row
-                            }
-                        } else {
-                            preferredResultBySubmissionID[key] = row
-                        }
-                    }
-
-                    for submission in submissions {
-                        guard let subID = submission.id,
-                            let gradePercent = bestPercentBySubmissionID[subID]
-                        else {
-                            continue
-                        }
-                        let setupID = submission.testSetupID
-                        let existing = bestGradePercentBySetupID[setupID] ?? 0
-                        if gradePercent > existing {
-                            bestGradePercentBySetupID[setupID] = gradePercent
-                        }
-                    }
-
-                    let disabledBySetup = try await disabledFetch
-                    let perSubBySetup = try await perSubFetch
-                    for (setupID, latest) in latestSubmissionBySetupID {
-                        guard let latestSubmission = grouped[setupID]?.first(where: { $0.id == latest.submissionID }),
-                            let result = preferredResultBySubmissionID[latest.submissionID],
-                            let collection = decodedCollection(from: result.collectionJSON),
-                            let gradePercent = gradePercent(from: collection)
-                        else {
-                            continue
-                        }
-                        let latestAttempt = latestSubmission.attemptNumber ?? 1
-                        let priorSub = grouped[setupID]?.first(where: { $0.attemptNumber == latestAttempt - 1 })
-                        let priorGradePercent: Int? = priorSub.flatMap { ps in
-                            guard let psID = ps.id,
-                                let pr = preferredResultBySubmissionID[psID]
-                            else { return nil }
-                            return pr.gradePercentValue
-                        }
-                        latestBadgesBySetupID[setupID] = AchievementBadge.forSubmission(
-                            BadgeContext(
-                                attemptNumber: latestAttempt,
-                                gradePercent: gradePercent,
-                                executionTimeMs: collection.executionTimeMs,
-                                priorGradePercent: priorGradePercent
-                            ),
-                            achievements: perSubBySetup[setupID],
-                            disabled: disabledBySetup[setupID] ?? [])
-                    }
-
-                    let classAchievements = try await classAchievementsFetch
-                    for ach in classAchievements {
-                        if let badge = AchievementBadge.forClassAchievement(
-                            ach.achievementID, disabled: disabledBySetup[ach.testSetupID] ?? [])
-                        {
-                            latestBadgesBySetupID[ach.testSetupID, default: []].append(badge)
-                        }
-                    }
-                }
-            }
-        }
+        let gradeData = try await Self.loadStudentDashboardGradeData(
+            req: req, user: user, setups: setups, fmt: fmt)
 
         let sortedSetups = setups.sorted { lhs, rhs in
             let lhsID = lhs.id ?? ""
             let rhsID = rhs.id ?? ""
-            let lhsOrder = assignmentBySetup[lhsID]?.sortOrder
-            let rhsOrder = assignmentBySetup[rhsID]?.sortOrder
-
-            switch (lhsOrder, rhsOrder) {
-            case (let l?, let r?) where l != r:
-                return l < r
-            default:
-                let lhsCreated = lhs.createdAt ?? .distantPast
-                let rhsCreated = rhs.createdAt ?? .distantPast
-                if lhsCreated != rhsCreated { return lhsCreated > rhsCreated }
-                return lhsID < rhsID
-            }
+            return assignmentDisplayOrderPrecedes(
+                lhsSortOrder: assignmentBySetup[lhsID]?.sortOrder,
+                rhsSortOrder: assignmentBySetup[rhsID]?.sortOrder,
+                lhsCreatedAt: lhs.createdAt, rhsCreatedAt: rhs.createdAt,
+                lhsSetupID: lhsID, rhsSetupID: rhsID)
         }
 
         // Notebook presence drives the Edit button.  The zip-derived answer
@@ -329,121 +181,17 @@ struct WebRoutes: RouteCollection {
             }
         }
 
-        let rows = sortedSetups.map { setup -> TestSetupRow in
-            let setupID = setup.id ?? ""
-            let data = Data(setup.manifest.utf8)
-            let props = decodeManifest(from: data)
-            let assignment = assignmentBySetup[setupID]
-            let latestSubmission = latestSubmissionBySetupID[setupID]
-            let submissionCount = submissionCountBySetupID[setupID] ?? 0
-            // A future open date drives the "Opens …" hint in the Due column,
-            // but not a distinct status — every assignment is scheduled, so a
-            // "scheduled" badge would add no signal.
-            let notYetOpen: Bool = {
-                guard let assignment, let startsAt = assignment.startsAt else { return false }
-                return Date() < startsAt
-            }()
-            // Preview is staff-only: staff see it functioning as "open" with a
-            // subtle staff-only marker, while to students it is indistinguishable
-            // from "closed". So the displayed status is resolved per viewer.
-            let status: String
-            let staffOnly: Bool
-            if let assignment {
-                switch assignment.visibility {
-                case .open:
-                    status = "open"
-                    staffOnly = false
-                case .closed:
-                    status = "closed"
-                    staffOnly = false
-                case .preview:
-                    status = isActiveCourseStaff ? "open" : "closed"
-                    staffOnly = isActiveCourseStaff
-                }
-            } else {
-                status = "unpublished"
-                staffOnly = false
-            }
-            // True when the setup has a flat notebook file on disk, or the zip
-            // contains at least one .ipynb entry (resolved above via the cache).
-            let hasNotebook = hasNotebookBySetupID[setupID] ?? false
-            let vanityBaseURL: String? = {
-                guard let assignment,
-                    let courseCode = courseState.active?.code,
-                    !courseCode.isEmpty,
-                    !assignment.slug.isEmpty
-                else { return nil }
-                return VanityURLRoutes.vanityPath(courseCode: courseCode, assignmentSlug: assignment.slug)
-            }()
-            // Active extension for this student on this assignment.  Drives
-            // the Submit button and Due column when the assignment-wide
-            // deadline has passed but this user retains submit privileges.
-            let extensionDueAt = extensionDueAtBySetupID[setupID]
-            let baselineDueAt = assignment?.dueAt
-            let hasActiveExtension = studentHasActiveExtension(extensionDueAt: extensionDueAt)
-            let effectiveDueAt = laterDeadline(
-                baseline: baselineDueAt, extensionDueAt: extensionDueAt)
-            let isOpenForThisUser: Bool = {
-                guard let assignment else { return false }
-                // Preview is open for staff, closed for students; staff testing a
-                // preview also bypass the future-open-date gate (see submissionGate).
-                let gate = assignment.visibility.submissionGate(isStaff: isActiveCourseStaff)
-                return isAssignmentOpenForUser(
-                    isOpen: gate.treatAsOpen,
-                    overrideActive: assignment.deadlineOverrideActive ?? false,
-                    baselineDueAt: baselineDueAt,
-                    effectiveDueAt: effectiveDueAt,
-                    hasActiveExtension: hasActiveExtension,
-                    startsAt: gate.honorsStartDate ? assignment.startsAt : nil
-                )
-            }()
-            // A published-but-closed assignment is openable read-only, so it
-            // still gets the open-notebook action even for a student who never
-            // engaged with it (the page renders read-only and hides Submit).
-            let canEdit =
-                isOpenForThisUser
-                || previouslyOpenedSetupIDs.contains(setupID)
-                || (assignment.map { assignmentVisibleToStudentByState($0) } ?? false)
-            // An active per-student extension keeps a class-closed assignment
-            // open for this one student, so the dashboard badge should read as
-            // actionable ("extended") rather than the misleading class-wide
-            // "closed" — the more so on phones, where the due column (with its
-            // "(extension)" note) is hidden and the badge is the only status
-            // signal. Scoped to the genuine published-then-closed case; preview /
-            // unpublished are untouched, and staff never carry extensions.
-            let displayStatus = (hasActiveExtension && status == "closed") ? "extended" : status
-            let badgeSplit = AchievementBadge.dashboardSplit(latestBadgesBySetupID[setupID] ?? [])
-            return TestSetupRow(
-                id: setupID,
-                title: assignment?.title,
-                notebookURL: vanityBaseURL ?? "/testsetups/\(setupID)/notebook",
-                submitURL: vanityBaseURL.map { "\($0)/submit" } ?? "/testsetups/\(setupID)/submit",
-                historyURL: vanityBaseURL.map { "\($0)/history" } ?? "/testsetups/\(setupID)/history",
-                suiteCount: props?.testSuites.count ?? 0,
-                createdAt: setup.createdAt.map { fmt.string(from: $0) } ?? "—",
-                dueAt: assignment?.dueAt.map { fmt.string(from: $0) },
-                opensAtText: notYetOpen ? assignment?.startsAt.map { fmt.string(from: $0) } : nil,
-                status: displayStatus,
-                staffOnly: staffOnly,
-                isOpen: isOpenForThisUser,
-                canEdit: canEdit,
-                gradingMode: props?.gradingMode.rawValue ?? GradingMode.worker.rawValue,
-                hasNotebook: hasNotebook,
-                submissionCount: submissionCount,
-                hasLatestSubmission: latestSubmission != nil,
-                latestSubmissionID: latestSubmission?.submissionID ?? "",
-                latestSubmittedAtText: latestSubmission?.submittedAtText ?? "—",
-                additionalSubmissionCount: max(submissionCount - 1, 0),
-                bestGradeText: overridePercentBySetupID[setupID].map { "\($0)%" }
-                    ?? bestGradePercentBySetupID[setupID].map { "\($0)%" },
-                gradeIsOverridden: overridePercentBySetupID[setupID] != nil,
-                badges: badgeSplit.visible,
-                extraBadgeCount: badgeSplit.extraCount,
-                extraBadgesTooltip: badgeSplit.extraTooltip,
-                hasActiveExtension: hasActiveExtension,
-                effectiveDueAtText: effectiveDueAt.map { fmt.string(from: $0) }
-            )
-        }
+        let rowContext = IndexRowContext(
+            fmt: fmt,
+            assignmentBySetup: assignmentBySetup,
+            gradeData: gradeData,
+            extensionDueAtBySetupID: extensionDueAtBySetupID,
+            previouslyOpenedSetupIDs: previouslyOpenedSetupIDs,
+            isActiveCourseStaff: isActiveCourseStaff,
+            activeCourseCode: courseState.active?.code,
+            hasNotebookBySetupID: hasNotebookBySetupID
+        )
+        let rows = sortedSetups.map { Self.buildTestSetupRow(setup: $0, context: rowContext) }
 
         // Sections for the active course (fetch started up top) enable the
         // grouped display below.
@@ -458,27 +206,17 @@ struct WebRoutes: RouteCollection {
             uniquingKeysWith: { first, _ in first }
         )
 
-        // Group rows by section; rows without a matching section → ungrouped.
-        var rowsBySectionID: [UUID: [TestSetupRow]] = [:]
-        var ungroupedSetups: [TestSetupRow] = []
-        for row in rows {
-            if let sID = sectionBySetupID[row.id] {
-                rowsBySectionID[sID, default: []].append(row)
-            } else {
-                ungroupedSetups.append(row)
-            }
-        }
-
-        // Build the ordered display groups: named sections (skipping any with no
-        // visible items) first, then a trailing unnamed bucket for ungrouped items.
-        var displayGroups: [IndexDisplayGroup] = allSections.compactMap { section in
-            guard let sID = section.id else { return nil }
-            let sectionRows = rowsBySectionID[sID] ?? []
-            guard !sectionRows.isEmpty else { return nil }
-            return IndexDisplayGroup(name: section.name, setups: sectionRows)
-        }
-        if !ungroupedSetups.isEmpty {
-            displayGroups.append(IndexDisplayGroup(name: nil, setups: ungroupedSetups))
+        // Ordered display groups (shared fold, #1118): named sections with
+        // visible items first, then a trailing unnamed bucket for ungrouped.
+        let grouped = groupRowsBySection(
+            rows: rows, sections: allSections, includeEmptySections: false,
+            sectionIDForRow: { sectionBySetupID[$0.id] },
+            makeSection: { section, sectionRows in
+                IndexDisplayGroup(name: section.name, setups: sectionRows)
+            })
+        var displayGroups = grouped.sections
+        if !grouped.ungrouped.isEmpty {
+            displayGroups.append(IndexDisplayGroup(name: nil, setups: grouped.ungrouped))
         }
 
         return try await req.view.render(
